refactorizing book-new and book-list
